chore(backend): create app from dependencies
